rearrange logical nesting
