+++ cmd/db.go
@@ -23,22 +23,22 @@ import (
 	"github.com/spf13/cobra"
 )
 
-// restoreCmd represents the restore command
-var restoreCmd = &cobra.Command{
-	Use:   "restore",
-	Short: "restore data from google cloud storage",
+// dbCmd represents the db command
+var dbCmd = &cobra.Command{
+	Use:   "db",
+	Short: "db-related commands",
 }
 
 func init() {
-	rootCmd.AddCommand(restoreCmd)
+	rootCmd.AddCommand(dbCmd)
 
 	// Here you will define your flags and configuration settings.
 
 	// Cobra supports Persistent Flags which will work for this command
 	// and all subcommands, e.g.:
-	// restoreCmd.PersistentFlags().String("foo", "", "A help for foo")
+	// dbCmd.PersistentFlags().String("foo", "", "A help for foo")
 
 	// Cobra supports local flags which will only run when this command
 	// is called directly, e.g.:
-	// restoreCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	// dbCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }
