+++ cmd/staking.go
@@ -20,57 +20,25 @@
 package cmd
 
 import (
-	"fmt"
-	"os"
-
-	"github.com/ava-labs/avalanchego/staking"
 	"github.com/spf13/cobra"
-
-	"github.com/patrick-ogrady/snowplow/pkg/utils"
 )
 
-// createCmd represents the create command
-var createCmd = &cobra.Command{
-	Use:   "create",
-	Short: "generate new staking credentials",
-	RunE:  createFunc,
+// stakingCmd represents the staking command
+var stakingCmd = &cobra.Command{
+	Use:   "staking",
+	Short: "staking-related commands",
 }
 
 func init() {
-	rootCmd.AddCommand(createCmd)
+	rootCmd.AddCommand(stakingCmd)
 
 	// Here you will define your flags and configuration settings.
 
 	// Cobra supports Persistent Flags which will work for this command
 	// and all subcommands, e.g.:
-	// createCmd.PersistentFlags().String("foo", "", "A help for foo")
+	// stakingCmd.PersistentFlags().String("foo", "", "A help for foo")
 
 	// Cobra supports local flags which will only run when this command
 	// is called directly, e.g.:
-	// createCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-}
-
-func createFunc(cmd *cobra.Command, args []string) error {
-	// Check if credentialDirectory is empty
-	if _, err := os.Stat(stakingDirectory); !os.IsNotExist(err) {
-		return fmt.Errorf("%s is not an empty directory", stakingDirectory)
-	}
-
-	// Generate Staking Key
-	if err := staking.GenerateStakingKeyCert(stakingKeyPath, stakingCertPath); err != nil {
-		return fmt.Errorf("%w: could not create staking credentials", err)
-	}
-
-	// Log NodeID
-	nodeID, err := utils.LoadNodeID(stakingCertPath)
-	if err != nil {
-		return fmt.Errorf("%w: could not calculate NodeID", err)
-	}
-	fmt.Printf(
-		"created new credentials for %s in %s\n",
-		utils.PrintableNodeID(nodeID),
-		stakingDirectory,
-	)
-
-	return nil
+	// stakingCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }
